% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_one_numeric}
\alias{check_one_numeric}
\title{check_one_numeric}
\arguments{
\item{value}{The parameter to test.}

\item{...}{Parameters passed out to the check_param_type_n_length() function}
}
\value{
\code{NULL}
}
\description{
Shortcut function to check_param_type_n_length(), to check a unique numeric value.
}
\seealso{
\link{check_param_type_n_length}

\link{check_parameter_type},\link{check_parameter_length}
}
\author{
L.Pavot
}
