% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_lines}
\alias{collapse_lines}
\title{collapse_lines - to paste strings with collapse = "\\n"}
\usage{
collapse_lines(..., sep = "\\n")
}
\arguments{
\item{...}{passed to \code{\link[base:paste]{base::paste0()}}}

\item{sep}{set the separator. Deafult is "\\n"}
}
\description{
collapse_lines() does exactly when paste does, but default collapse = "\\n"
}
\examples{

collapse_lines(list("a message ", "in multiple", "parts"))

}
\author{
L.Pavot
}
