% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_helpers.R
\name{df_force_numeric}
\alias{df_force_numeric}
\title{Convert data frame to numeric.}
\usage{
df_force_numeric(df, cols = NULL)
}
\arguments{
\item{df}{The data frame.}

\item{cols}{The set of columns to convert to numeric.
By default (when set to \code{NULL}) all integer columns are converted.
Set it to a character vector containing the names of the columns you
want to convert, or ton integer vector containing the indices of the
columns. Can be used to force conversion of non integer columns.}
}
\value{
The converted \code{data.frame}.
}
\description{
df_force_numeric
Converts integer columns of a data frame into numeric.
}
\examples{
# Convert an integer data frame
df <- data.frame(a = as.integer(c(1, 4)), b = as.integer(c(6, 5)))
df <- W4MRUtils::df_force_numeric(df)

}
