% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_datasets.R
\name{Filtering_datasets}
\alias{Filtering_datasets}
\title{Filtering tools for Galaxy scripting}
\description{
Filtering tools to reduce a dataset,
based on W4M 3-tables format
Coded by: M.Petera,

R functions to use in R scripts and wrappers
to make things easier

V0: script structure + first function
}
