% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{get_base_dir}
\alias{get_base_dir}
\title{get_base_dir - to get... the base directory}
\usage{
get_base_dir()
}
\value{
the directory path of the main script. PWD otherwise.
}
\description{
get_base_dir
}
\examples{
print(get_base_dir())

}
