% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_logger}
\alias{get_logger}
\title{Instantiate a Logger}
\usage{
get_logger(name, recreate = FALSE, ...)
}
\arguments{
\item{name}{the name of the logger to get or create. This name will
be used in logs to differentiate from which part of you program
comes which lines of log. See the example of usage bellow.}

\item{recreate}{logical=FALSE tells whether to recreate the logger
of the given name or not. Preferably, one should not recreate
a new logger each time.}

\item{...}{Parameters passed out to the W4MLogger() function}
}
\value{
A new W4MLogger instance if it did not exist or if recreate is
TRUE. Otherwise, a new W4MLogger instance.
}
\description{
Create a logger of the given name. You can call again \code{get_logger} and
provide the same name to get the same logger. It will not be recreated
unless recreate is TRUE.
}
\details{
get_logger
}
\author{
L.Pavot
}
