% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy.R
\name{get_r_env}
\alias{get_r_env}
\title{get_r_env - provides env vars begining with R_*}
\usage{
get_r_env()
}
\value{
a \code{list} of environment variables which
begin by \code{R_}.
}
\description{
Returns a list of env vars if the start with R_*.
}
\seealso{
\link{run_galaxy_processing}
}
\author{
L.Pavot
}
