% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy.R
\name{in_galaxy_env}
\alias{in_galaxy_env}
\title{in_galaxy_env - check if the script has been run by galaxy}
\usage{
in_galaxy_env()
}
\value{
A logical - whether the script has been run by galaxy or not.
}
\description{
\code{in_galaxy_env} returns \code{TRUE} if it detects some
galaxy-specific environment variables. \code{FALSE} otherwise.
}
