% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_checks.R
\name{intensity_checks}
\alias{intensity_checks}
\title{Intensity checking tools for Galaxy scripting}
\description{
Intensity checking tools to compute indices per variable,
based on the W4M 3-tables format
Original code of first function by A.Fernandes, updates by M.Petera

R functions to use in R scripts and wrappers
to make things easier

V0: script structure + first function (intens_check)
}
