% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcheckLibrary.R
\name{match2}
\alias{match2}
\title{Table match check functions}
\usage{
match2(data_matrix, metadata, metadata_type)
}
\arguments{
\item{data_matrix}{data.frame containing data_matrix}

\item{metadata}{data.frame containing sample_metadata or variable_metadata}

\item{metadata_type}{"sample" or "variable" depending on metadata content}
}
\value{
\code{character} vector a list of errors encountered
}
\description{
match2
To check if data_matrix and (variable or sample)metadata match regarding
identifiers
}
\author{
M.Petera
}
