% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcheckLibrary.R
\name{match3}
\alias{match3}
\title{match3}
\usage{
match3(data_matrix, sample_metadata, variable_metadata)
}
\arguments{
\item{data_matrix}{data.frame containing data_matrix}

\item{sample_metadata}{data.frame containing sample_metadata}

\item{variable_metadata}{data.frame containing variable_metadata}
}
\value{
\code{character} vector a list of errors encountered
}
\description{
match3
To check if the 3 standard tables match regarding identifiers
}
\author{
M.Petera
}
