% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{metab_merge}
\alias{metab_merge}
\title{Merging a dataMatrix with a metadata file}
\usage{
metab_merge(DM, meta, metype = "sample")
}
\arguments{
\item{DM}{Data.frame corresponding to the dataMatrix you want to merge with a metadata table}

\item{meta}{Data.frame corresponding to the metadata table you want to merge with the dataMatrix}

\item{metype}{Character string indicating the type of metadata table used; should be one between "sample" and "variable"}
}
\value{
a \code{data.frame} corresponding to a combined table with both intensities and metadata
}
\description{
metab_merge
Function to merge the dataMatrix table with one of its corresponding
metadata table (sampleMetadata or variableMetadata)
}
\author{
M.Petera
}
