% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{mini_tools}
\alias{mini_tools}
\title{Mini tools for Galaxy scripting}
\description{
Mini tools for Galaxy scripting
Mini tools for Galaxy scripting
Coded by: M.Petera,

R functions to use in R scripts and wrappers
to make things easier (lightening code, reducing verbose...)

V0: script structure + first functions
V1: addition of functions to handle special characters in identifiers
}
