% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optparse_helper.R
\name{optparse_character}
\alias{optparse_character}
\title{optparse_character - define a command parameter as string}
\usage{
optparse_character(help = "No documentation yet.", short = NULL, default = 0)
}
\arguments{
\item{help}{\itemize{
\item The help string to display when --help is triggered
}}

\item{short}{\itemize{
\item The shortcut fir this parameter. For example
for a --output param, we could use
\code{optparse_flag(short = "-o", ...)} to set the "-o" shortcut.
}}

\item{default}{\itemize{
\item The default value this parameter will hold.
}}
}
\description{
To be used with \code{optparse_parameters}. This function tells
the provided parameter is to be parsed as a single string.
}
\examples{

str(optparse_parameters(
  a_parameter = optparse_character(),
  args = list("--a-parameter", "42")
))

}
\seealso{
\code{\link[=optparse_parameters]{optparse_parameters()}}
}
\author{
L.Pavot
}
