% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{parse_command_args_tf}
\alias{parse_command_args_tf}
\title{Parse Command Args True/False}
\usage{
parse_command_args_tf(...)
}
\description{
parse_command_args_tf
Function to replace the default batch::parseCommandArgs to
solve an issue with batch if arguments are logical TRUE/FALSE
}
\examples{
\dontrun{
#interpretation of arguments given in command line as an R list of objects
args <- parse_command_args_tf(evaluate = FALSE)
}

}
\author{
M.Petera
}
