% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printf}
\alias{printf}
\title{printf - to format a string and print it}
\usage{
printf(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:sprintf]{base::sprintf}}
  \describe{
    \item{\code{fmt}}{a character vector of format strings, each of up to 8192 bytes.}
  }}
}
\description{
printf calls sprintf of its parameters to build the error message
and prints with the given message
}
\examples{

file <- "/tmp/test"
printf("Error in file: ", file)

}
\author{
L.Pavot
}
