% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcheckLibrary.R
\name{R Check Tools}
\alias{R Check Tools}
\title{R check Tools}
\description{
R functions to use in R scripts
(management of various generic subroutines)

V0: script structure + first functions
V1: More detailed error messages in match functions
}
\author{
M.Petera,
}
