% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy.R
\name{run_galaxy_function}
\alias{run_galaxy_function}
\title{run_galaxy_function - automate running functions in galaxy}
\usage{
run_galaxy_function(tool_name, func, ...)
}
\arguments{
\item{tool_name}{\itemize{
\item The name of the tool
}}

\item{func}{\itemize{
\item The function to be run, after galaxy header is displayed
}}

\item{...}{\itemize{
\item Parameters propagated to \link{run_galaxy_processing}
}}
}
\description{
This function executes the provided function as a galaxy processing
This provided function is expected to take two parameters:
\itemize{
\item args, a list of command line parameters
\item logger, the logger created for the tool
}
}
\seealso{
\link{run_galaxy_processing}
}
\author{
L.Pavot
}
