% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{source_local}
\alias{source_local}
\title{source_local - source file, from absolute or relative path}
\usage{
source_local(..., env = FALSE, do_print = FALSE, keep_source = TRUE)
}
\arguments{
\item{...}{paths, character vector of file paths to source}

\item{env}{an environement in which to source the paths}

\item{do_print}{a logical, telling whether to print sourced paths or
not}

\item{keep_source}{See the parameter keep.source from source}
}
\value{
a vector resulting from the sourcing of the files provided.
}
\description{
source_local
Transforms a relative path to an absolute one, and sources the path.
This helps source files located relatively to the main script without
the need to know from where it was run.
}
\seealso{
\code{\link[=source]{source()}}
}
