% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopf}
\alias{stopf}
\title{stopf - to stop and format message}
\usage{
stopf(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[base:sprintf]{base::sprintf}}
  \describe{
    \item{\code{fmt}}{a character vector of format strings, each of up to 8192 bytes.}
  }}
}
\description{
stopf calls sprintf of its parameters to build the error message
and stops with the given message
}
\examples{

tryCatch({
  file <- "/tmp/test"
  stopf("Error in \%s file.", file)
}, error = function(error) {
  print(error)
})

}
\author{
L.Pavot
}
