% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy.R
\name{unmangle_galaxy_param}
\alias{unmangle_galaxy_param}
\title{unmangle_galaxy_param - revert effects of galaxy manglings.}
\usage{
unmangle_galaxy_param(args)
}
\arguments{
\item{args}{named \code{list} - contains params_name=value.}
}
\value{
a named \code{list} - with unmangled parameter name and
values.
}
\description{
When running a tool from galaxy, the command line may be
altered because some forbidden chars have been translated by galaxy.

This function takes \code{args} are invert the galaxy's mangling process.
}
\seealso{
\link{run_galaxy_processing}, \link{unmangle_galaxy_string}
}
\author{
L.Pavot
}
