% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxy.R
\name{unmangle_galaxy_string}
\alias{unmangle_galaxy_string}
\title{unmangle_galaxy_string - revert effects of galaxy mangling}
\usage{
unmangle_galaxy_string(string)
}
\arguments{
\item{string}{\itemize{
\item the character vector to fix mangling.
}}
}
\value{
string - the character vector, fixed.
}
\description{
Revert effect of string mangling from galaxy on the given string.
}
\seealso{
\link{run_galaxy_processing}, \link{unmangle_galaxy_param}
}
\author{
L.Pavot
}
