% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_[info,warning,error,debug,verbose]}
\alias{W4MLogger_[info,warning,error,debug,verbose]}
\alias{W4MLogger$info}
\alias{W4MLogger$warning}
\alias{W4MLogger$error}
\alias{W4MLogger$debug}
\alias{W4MLogger$verbose}
\title{Log info/warning/error/debug/verbose messages}
\description{
Call one of the following function when you want a message to be printed
or written in a log file:
\itemize{
\item \code{your_logger$info("A info message")} ;
\item \code{your_logger$warning("A warning message")} ;
\item \code{your_logger$error("A error message")} ;
\item \code{your_logger$debug("A debug message")} ;
\item \code{your_logger$verbose.("A verbose. message")}
}

If the corresponding level is activated (with your_logger$set_info(TRUE),
your_logger$set_debug(TRUE), etc...), these functions will print the
message provided in the terminal and in logs files, if there were some
provided at the creation of the logger.

If the corresponding log level is deactivated, these function will
not do anything. So, do not hesitate to use them a lot, and activate them
when needed.

See \link{get_logger} for example usages.
}
\author{
L.Pavot
}
