% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_finalize}
\alias{W4MLogger_finalize}
\title{W4MLogger_finalize}
\description{
The function W4MLogger$finalize is the destructor function of this
class. It closes every files that was opened by the logger, or that
was provided during execution. It has to be considered internal.

The function W4MLogger$finalize is the destructor function of this
class. It closes every files that was opened by the logger, or that
was provided during execution. It has to be considered internal.

The function W4MLogger$finalize is the destructor function of this
class. It closes every files that was opened by the logger, or that
was provided during execution. It has to be considered internal.

The function W4MLogger$finalize is the destructor function of this
class. It closes every files that was opened by the logger, or that
was provided during execution. It has to be considered internal.

The function W4MLogger$finalize is the destructor function of this
class. It closes every files that was opened by the logger, or that
was provided during execution. It has to be considered internal.
}
