% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_set_debug}
\alias{W4MLogger_set_debug}
\title{W4MLogger$set_debug}
\arguments{
\item{value}{logical TRUE/FALSE to activate/deactivate debug logging}

\item{default}{logical set to TRUE to use debug by default}
}
\value{
.self the current W4MLogger instance
}
\description{
This method activate or deactivate the logging of debugs messages
}
