% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_set_error}
\alias{W4MLogger_set_error}
\title{W4MLogger$set_error}
\arguments{
\item{value}{logical TRUE/FALSE to activate/deactivate error logging}

\item{default}{logical set to TRUE to use error by default}
}
\value{
.self the current W4MLogger instance
}
\description{
This method activate or deactivate the logging of errors messages
}
