% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_set_info}
\alias{W4MLogger_set_info}
\title{W4MLogger$set_info}
\arguments{
\item{value}{logical TRUE/FALSE to activate/deactivate info logging}

\item{default}{logical set to TRUE to use info by default}
}
\value{
.self the current W4MLogger instance
}
\description{
This method activate or deactivate the logging of info messages
}
