% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_set_verbose}
\alias{W4MLogger_set_verbose}
\title{W4MLogger$set_verbose}
\arguments{
\item{value}{logical TRUE/FALSE to activate/deactivate verbose logging}

\item{default}{logical set to TRUE to use verbose by default}
}
\value{
.self the current W4MLogger instance
}
\description{
This method activate or deactivate the logging of verbose messages
}
