% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{W4MLogger_set_warning}
\alias{W4MLogger_set_warning}
\title{W4MLogger$set_warning}
\arguments{
\item{value}{logical TRUE/FALSE to activate/deactivate warning logging}

\item{default}{logical set to TRUE to use warning by default}
}
\value{
.self the current W4MLogger instance
}
\description{
This method activate or deactivate the logging of warnings messages
}
