% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcheckLibrary.R
\name{check_err}
\alias{check_err}
\title{Check Errors}
\usage{
check_err(err_stock)
}
\arguments{
\item{err_stock}{vector of results returned by check functions}
}
\value{
\code{NULL}
}
\description{
check_err
Generic function stop in error if problems have been encountered
}
\author{
M.Petera
}
