% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{reproduce_id}
\alias{reproduce_id}
\title{Reproduce ID}
\usage{
reproduce_id(data_matrix, metadata, metadata_type, id_match)
}
\arguments{
\item{data_matrix}{data.frame containing data_matrix}

\item{metadata}{data.frame containing samplemetadata or variablemetadata}

\item{metadata_type}{"sample" or "variable" depending on metadata content}

\item{id_match}{'id_match' element produced by stock_id}
}
\description{
reproduce_id
reproduce_id reinjects original identifiers and original order into
final tables
}
\examples{
\dontrun{
B <- reproduce_id(datamatrix, sample_metadata, "sample", A)
datamatrix <- B$data_matrix
sample_metadata <- B$metadata
}

}
\author{
M.Petera
}
