% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{source_local}
\alias{source_local}
\title{source local}
\usage{
source_local(...)
}
\arguments{
\item{...}{paths, character vector of file paths to source}
}
\value{
a vector of
}
\description{
source_local
Transforms a path to be relative to the main script, and sources the path.
This helps source files located relatively to the main script without
the need to know from where it was run.
}
\examples{
\donttest{
   
   W4MRUtils::source_local("example.R","RcheckLibrary.R")
}
}
