% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{stock_id}
\alias{stock_id}
\title{Stock ID}
\usage{
stock_id(data_matrix, metadata, metadata_type)
}
\arguments{
\item{data_matrix}{data.frame containing data_matrix}

\item{metadata}{data.frame containing samplemetadata or variablemetadata}

\item{metadata_type}{"sample" or "variable" depending on metadata content}
}
\description{
stock_id
Functions to stock identifiers before applying make.names() and
to reinject it into final matrices.
stock_id stocks original identifiers and original order
needs checked data regarding table match.
Note: it reproduces the original order of datasets' identifiers
Function to be used at the very end, when exporting tables
}
\examples{
\dontrun{
A <- stock_id(myDM, mysM, "sample")
myDM <- A$data_matrix
mysM <- A$metadata
A <- A$id_match
}

}
\author{
M.Petera
}
