% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniTools.R
\name{stock_id}
\alias{stock_id}
\title{Stock ID}
\usage{
stock_id(data_matrix, metadata, metadata_type)
}
\arguments{
\item{stock_id:}{stocks original identifiers and original order
needs checked data regarding table match}

\item{reproduce_id:}{reinjects original identifiers and original order into
final tables}
}
\description{
stock_id
Functions to stock identifiers before applying make.names() and
to reinject it into final matrices
Note: it reproduces the original order of datasets' identifiers
Function to be used at the very end, when exporting tables
}
\examples{
\dontrun{
A <- stock_id(myDM, mysM, "sample")
myDM <- A$data_matrix
mysM <- A$metadata
A <- A$id_match
}

}
\author{
M.Petera
}
